work until
Precomputed inverted indexes allow the documents that contain any particular query term to be quickly identified. If the objective is to compute a similarity score for every document that contains any of the query terms, then exhaustive processing of the inverted list associated with each query term is sufficient. Two alternatives have been explored: term-at-a-time (TAAT) processing, where inverted lists are opened sequentially, and a state vector of partially computed similarity scores is maintained; and document-at-a-time (DAAT) processing,
in which the inverted lists are opened concurrently, and each
document is fully scored before any other document is considered.
Document-at-a-time processing requires that the inverted lists be
stored in document-sorted order; term-at-a-time allows other index
orderings.